Token authentication
Implemented token attachment on user login on the server.
Added logic for displaying different navbar if user is logged in or not.

Added 'hello username' in the NavbarUserMenu.
Changed the logic on register -> User doesn't get automatically logged
in on register.
Added logic on logout - remove token.
Fixed some form messages, but more messages need fixing.